refactor: streaming methods fall in line
