+++ tests/test_streams.py
@@ -1,7 +1,5 @@
-import uuid
-import os
 import random
-from xenon.oop import FileSystem
+from xenon.objects import FileSystem
 
 
 def read_lines(stream):
